requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0

SQLAlchemy==2.0.36
psycopg2-binary==2.9.10

pydantic==2.10.4
pydantic-settings==2.7.0

PyJWT==2.10.1
alembic==1.14.0

passlib[bcrypt]==1.7.4
email-validator==2.1.1
email-validator==2.1.1
